tests/wpt/web-platform-tests/css/css-contain/container-queries/container-name-tree-scoped.html

<!doctype html>
<meta charset="utf-8">
<title>CSS Container Queries Test: Tree scoped container-name</title>
<link rel="help" href="https://drafts.csswg.org/css-contain-3/#query-container">
<link rel="help" href="https://drafts.csswg.org/css-scoping-1/#shadow-names">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/declarative-shadow-dom-polyfill.js"></script>
<script src="support/cq-testcommon.js"></script>

<div id="container-name-host">
  <div>
    <template shadowroot="open">
      <style>
        :host { container-name: foo; }
      </style>
      <slot></slot>
    </template>
    <div id="t1"></div>
  </div>
  <style>
    #container-name-host > div {
      container-type: inline-size;
    }
    #t1 { color: green; }
    @container foo (width > 0px) {
      #t1 { color: red; }
    }
  </style>
</div>

<div id="container-name-slotted">
  <div>
    <template shadowroot="open">
      <style>
        ::slotted(div) {
          container-name: foo;
        }
      </style>
      <slot></slot>
    </template>
    <div>
      <div id="t2"></div>
    </div>
  </div>
  <style>
    #container-name-slotted > div > div {
      container-type: inline-size;
    }
    #t2 { color: green; }
    @container foo (width > 0px) {
      #t2 { color: red; }
    }
  </style>
</div>

<script>
  setup(() => {
    assert_implements_container_queries();
    polyfill_declarative_shadow_dom(document);
  });

  const green = "rgb(0, 128, 0)";

  test(() => {
    assert_equals(getComputedStyle(t1).color, green);
  }, "Outer scope query should not match container-name set by :host rule in shadow tree");

  test(() => {
    assert_equals(getComputedStyle(t2).color, green);
  }, "Outer scope query should not match container-name set by ::slotted rule in shadow tree");

</script>
