third_party/blink/web_tests/external/wpt/css/css-contain/container-queries/container-name-tree-scoped-expected.txt

This is a testharness.js-based test.
FAIL Outer scope query should not match container-name set by :host rule in shadow tree assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
FAIL Outer scope query should not match container-name set by ::slotted rule in shadow tree assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
Harness: the test ran to completion.
